+++ valkey_glide_cluster.c
@@ -989,6 +989,14 @@ BGREWRITEAOF_METHOD_IMPL(ValkeyGlideCluster)
  * [, bool copy, bool replace, mixed credentials]) */
 MIGRATE_METHOD_IMPL(ValkeyGlideCluster)
 /* }}} */
+  
+/* {{{ proto ValkeyGlideCluster::reset() */
+RESET_METHOD_IMPL(ValkeyGlideCluster)
+/* }}} */
+
+/* {{{ proto ValkeyGlideCluster::save(mixed route) */
+SAVE_METHOD_IMPL(ValkeyGlideCluster)
+/* }}} */
 
 /* {{{ proto ValkeyGlideCluster::dbsize(string key)
  *     proto ValkeyGlideCluster::dbsize(array host_port) */