+++ valkey_glide_cluster.stub.php
@@ -590,6 +590,11 @@ class ValkeyGlideCluster
     public function bgRewriteAof(mixed $route): ValkeyGlideCluster|array|bool|string;
 
     /**
+     * @see ValkeyGlide::save
+     */
+    public function save(mixed $route): ValkeyGlideCluster|bool|string;
+
+    /**
      * @see ValkeyGlide::geoadd
      */
     public function geoadd(string $key, float $lng, float $lat, string $member, mixed ...$other_triples_and_options): ValkeyGlideCluster|int|false;
@@ -970,6 +975,11 @@ class ValkeyGlideCluster
      *               if one was.
      */
     public function ping(mixed $route, ?string $message = null): mixed;
+
+    /**
+     * @see ValkeyGlide::reset
+     */
+    public function reset(): ValkeyGlideCluster|bool|string;
 
     /**
      * @see ValkeyGlide::psetex